Fix reordering of portlets.
In some situations, the ISolgemaPortletManagerRetriever would not
include the portlet being moved, and the operation would fail with an
IndexError.

This commit adds a check for the portlet hash inclusion in the list
returned by the retriever, and uses the plain delta as the position if
it isn't included in the list.